requirements.txt
boto3==1.20.14
botocore==1.23.14
certifi==2021.10.8
charset-normalizer==2.0.8
deepdiff==5.6.0
idna==3.3
jmespath==0.10.0
ordered-set==4.0.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
